Fix compile error, trying to force a json return type
Not sure if this will fix anything; It's the last difference between my local repo and master, and I know angular tries to attach the xsrf token to a _csrf field in its requests, or in a dedicated 'X-CSRF-TOKEN' header if the request format is json. We need a x-xsrf-token header, and there isn't one. Ergo, perhaps the content-type isn't being set. But this change would only affect the log in request's response..